+++ pipeline/sig_verifier_mgmt.go
@@ -83,7 +83,7 @@ func (m *sigVerifierMgr) startBlockReceiverRoutine() {
 			case b := <-m.inputChan:
 				v.sendCh <- b
 				if m.metrics.Enabled {
-					m.metrics.SigVerifierMgrInTxs.Add(float64(len(b.Txs)))
+					m.metrics.SigVerifierMgrInTxs.Add(len(b.Txs))
 					m.metrics.SigVerifierMgrInputChLength.Set(len(m.inputChan))
 				}
 			}
@@ -128,7 +128,7 @@ func (m *sigVerifierMgr) startOutputWriterRoutine() {
 					}
 				}
 				if m.metrics.Enabled {
-					m.metrics.SigVerifierMgrOutTxs.Add(float64(len(responseBatch.Responses)))
+					m.metrics.SigVerifierMgrOutTxs.Add(len(responseBatch.Responses))
 				}
 			}
 		}
